fix: rename routingUtils to reactUtils
