+++ js/apps/system/_admin/aardvark/APP/react/src/App.js
@@ -25,7 +25,7 @@ require('./views/query/QueryReactView');
 require('./views/shards/distribution/ShardDistributionReactView');
 require('./views/services/ServicesReactView');
 
-require('./routingUtils');
+require('./reactUtils');
 
 // old libraries
 const jQuery = require('jquery');
